requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.2.1
Django==1.11.7
djangorestframework==3.7.3
djangorestframework-camel-case==0.2.0
enum34==1.1.6
funcy==1.10
ipython==5.5.0
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
